Split extension specs for rails 3/4
